+++ src/lib/utils.js
@@ -8,6 +8,19 @@ const utils = {
     return s.charAt(0).toUpperCase() + s.slice(1);
   },
 
+  /**
+   * Async for each loop
+   *
+   * @param  {array} array - Array to loop through
+   * @param  {function} callback - Function to apply to each array item loop
+   */
+  asyncForEach: async (array, callback) => {
+    for (let index = 0; index < array.length; index += 1) {
+      // eslint-disable-next-line callback-return, no-await-in-loop
+      await callback(array[index], index, array);
+    }
+  },
+
   removeCompendiumLinks: (text) => {
     const linkRegExTag = /@\w+\[(.*)\](\{.*?\})/g;
     const linkRegExNoTag = /@\w+\[(.*)\]/g;
@@ -279,27 +292,6 @@ const utils = {
     return undefined;
   },
 
-  getOrCreateFolder: async (root, entityType, folderName, folderColor = "") => {
-    let folder = game.folders.contents.find((f) =>
-      f.type === entityType && f.name === folderName
-      // if a root folder we want to match the root id for the parent folder
-      && (root ? root.id : null) === (f.folder?.id ?? null)
-    );
-    // console.warn(`Looking for ${root} ${entityType} ${folderName}`);
-    // console.warn(folder);
-    if (folder) return folder;
-    folder = await Folder.create(
-      {
-        name: folderName,
-        type: entityType,
-        color: folderColor,
-        parent: (root) ? root.id : null,
-      },
-      { displaySheet: false }
-    );
-    return folder;
-  },
-
   entityMap: () => {
     let entityTypes = new Map();
     entityTypes.set("spell", "Item");
@@ -328,22 +320,6 @@ const utils = {
     entityTypes.set("compendium", "Compendium");
     entityTypes.set("class", "Class");
     return entityTypes;
-  },
-
-  // eslint-disable-next-line no-unused-vars
-  getFolder: async (kind, subFolder = "", baseFolderName = "D&D Beyond Import", baseColor = "#6f0006", subColor = "#98020a", typeFolder = true) => {
-    const entityTypes = utils.entityMap();
-    const folderName = game.i18n.localize(`ddb-importer.item-type.${kind}`);
-    const entityType = entityTypes.get(kind);
-    const baseFolder = await utils.getOrCreateFolder(null, entityType, baseFolderName, baseColor);
-    const entityFolder = typeFolder ? await utils.getOrCreateFolder(baseFolder, entityType, folderName, subColor) : baseFolder;
-    if (subFolder !== "") {
-      const subFolderName = subFolder.charAt(0).toUpperCase() + subFolder.slice(1);
-      const typeFolder = await utils.getOrCreateFolder(entityFolder, entityType, subFolderName, subColor);
-      return typeFolder;
-    } else {
-      return entityFolder;
-    }
   },
 
   versionCompare: (v1, v2, options) => {